+++ CHANGELOG.md
@@ -7,6 +7,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased] - YYYY-MM-DD
 ### ➕ Additions
 * sp-enhancements: Add type definition to `addedEntries` argument for `getAllDependenciesForSpEntry` and `getAllEntriesThatDependOnSpEntry`
+* datamine-types: Add definitions for the following types
+	* `IEvolutionMaterial`
+	* `IUnit.evolution`
 
 ## [0.4.1] - 2020-04-12
 ### ➕ Additions